auth_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.3
pydantic==2.7.1
